Switch GPD, GEO and CARMA importers to ISO alpha 2 country codes
To use this, `config['target_countries']` must now be a list of
two-letter country codes instead of country names.

The rationale is that for Africa, full country names are
inconsistent (language, spelling, etc.), and so we need ISO country
codes to be able to match up different countries between databases.